refactor(cardMapping): 모든 멤버가 static 인 레코드를 레코드와 유틸로 분리, 객체로 풀어내기
