+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -5,19 +5,19 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 public class Dealer extends Participant {
-
+    
     private static final String DEALER_NICKNAME = "딜러";
     private static final int DEALER_SCORE = 16;
-
+    
     public Dealer() {
         super(DEALER_NICKNAME);
     }
-
+    
     @Override
     public boolean isDrawable() {
         return hand.getTotalScore() <= DEALER_SCORE;
     }
-
+    
     public Map<GameResult, Long> determineGameResult(List<Player> players) {
         return players.stream()
                 .map(this::determineGameResult)
@@ -25,7 +25,7 @@ public class Dealer extends Participant {
                         result -> result, Collectors.counting()
                 ));
     }
-
+    
     public GameResult determineGameResult(Player player) {
         if (isBusted()) {
             return GameResult.LOSE;
@@ -35,7 +35,7 @@ public class Dealer extends Participant {
         }
         return compareScore(getScore(), player.getScore());
     }
-
+    
     private GameResult compareScore(int score, int playerScore) {
         if (score == playerScore) {
             return GameResult.DRAW;
