+++ src/main/java/blackjack/view/dto/ParticipantHandScore.java
@@ -1,6 +1,7 @@
 package blackjack.view.dto;
 
 import blackjack.domain.participant.Participant;
+import blackjack.view.CardMapper;
 import java.util.List;
 
 public record ParticipantHandScore(
@@ -25,7 +26,7 @@ public record ParticipantHandScore(
     
     private static List<String> toCardDisplayNames(Participant participant) {
         return participant.getCards().stream()
-                .map(CardDisplayName::from)
+                .map(CardMapper::toDisplayName)
                 .map(CardDisplayName::displayName)
                 .toList();
     }
